pyproject.toml
[tool.poetry]
name = "datavis2d"
version = "0.1.0"
description = ""
authors = ["tocom <tcom242242@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.4.0"
torchvision = "^0.19.0"
matplotlib = "^3.9.2"
scikit-learn = "^1.5.1"
umap-learn = "^0.5.6"
pillow = "^10.4.0"
shap = "^0.46.0"
numpy = "1.25.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
